LayoutTests/platform/ios-simulator/fast/events/platform-wheelevent-paging-x-in-scrolling-div-expected.txt

PASS div.scrollTop is window.expectedScrollTop
FAIL div.scrollLeft should be 322. Was 0.
